Moved bar graph into its own component, added alerts to checkbox if you are not assigned to the task
