+++ pages/Ecosystem/[id].js
@@ -22,6 +22,7 @@ import { setDoc, doc, deleteDoc } from 'firebase/firestore';
 import Tab from '@mui/material/Tab';
 import Tabs from '@mui/material/Tabs';
 import PropTypes from 'prop-types';
+import BarGraph from '../../Components/BarGraph';
 
 export default function ecosystem() {
   const [addTask, setAddTasK] = useState(false);
@@ -189,7 +190,7 @@ export default function ecosystem() {
                         return (
                           <div className='flex' key={idx}>
                             {task.completed ? (
-                              task.userId === user.uid ? (
+                              task.assignedTo === user.uid ? (
                                 <CheckBoxIcon
                                   className='flex justify-end mr-3'
                                   onClick={() =>
@@ -197,9 +198,16 @@ export default function ecosystem() {
                                   }
                                 />
                               ) : (
-                                <CheckBoxIcon className='flex justify-end mr-3' />
+                                <CheckBoxIcon
+                                  onClick={() =>
+                                    alert(
+                                      'You cannot mark a task that is not assigned to you!'
+                                    )
+                                  }
+                                  className='flex justify-end mr-3'
+                                />
                               )
-                            ) : task.userId === user.uid ? (
+                            ) : task.assignedTo === user.uid ? (
                               <CheckBoxOutlineBlankIcon
                                 className='flex justify-end mr-3'
                                 onClick={() =>
@@ -207,7 +215,14 @@ export default function ecosystem() {
                                 }
                               />
                             ) : (
-                              <CheckBoxOutlineBlankIcon className='flex justify-end mr-3' />
+                              <CheckBoxOutlineBlankIcon
+                                onClick={() =>
+                                  alert(
+                                    'You cannot mark a task that is not assigned to you!'
+                                  )
+                                }
+                                className='flex justify-end mr-3'
+                              />
                             )}
                             <li key={idx} className='text-left p-1 ml-2'>
                               {task.name}
@@ -245,48 +260,10 @@ export default function ecosystem() {
         </div>
         <div className='flex h-1/2 w-full'>
           <div className='flex border border-black rounded-3xl justify-center w-full m-4'>
-            <Chart className='w-full' />
+            <BarGraph className='w-full' />
           </div>
         </div>
       </div>
     </>
   );
 }
-
-import {
-  Chart as ChartJS,
-  ArcElement,
-  Tooltip,
-  Legend,
-  registerables,
-} from 'chart.js';
-import { Bar } from 'react-chartjs-2';
-
-ChartJS.register(...registerables, Tooltip, Legend);
-
-const Chart = () => {
-  const data = {
-    labels: ['Scott', 'Mike', 'Cadre', 'Tasdid', 'Allan', 'Sarah', 'Emily'],
-    datasets: [
-      {
-        barPercentage: 0.5,
-        barThickness: 50,
-        maxBarThickness: 800,
-        minBarLength: 2,
-        data: [1, 6, 7, 4, 5, 2, 3],
-        backgroundColor: [
-          'rgb(255, 99, 132)',
-          'rgb(54, 162, 235)',
-          'rgb(170, 239, 139)',
-          'rgb(137, 167, 178)',
-          'rgb(107, 32, 173)',
-          'rgb(244, 175, 24)',
-          'rgb(204, 38, 26)',
-        ],
-        hoverOffset: 4,
-      },
-    ],
-  };
-
-  return <Bar data={data} />;
-};
